terminfo: Remove `create_function` dependency
This removes the use of `eval` from the Cli\TermInfo class. The new
compiler uses a lambda calculus approach with the use of anonymous
functions, which should perform slightly better than an interpreter,
albeit it is likely more difficult to read than the previous version.

It makes the PHP development team happy by removing the
DeprecationWarning emitted from the use of `create_function`.

Performance wise, the previous version performed at about 3.5us to
compile the `setaf` capability and 0.36us to execute it. The new version
performs at about 5.7us to compile the `setaf` capability and 0.57us to
execute it. So it's considerably slower but in the same performance
ballpark as the previous.